Enable some features found on maps without using the map author's code.  This lets us use those features for maps that require us to override the main *.gsc file to get the map to work in RotU, and when we can't decompile the map's fast file to use the map author's own function calls.  The map map_fart_house_v2 is an example of this case.
We can load glide pads, which move a player from point to point to point over time.
We can load teleporters that teleport a player from the trigger location to some destination.
We can load elevators, both vertical, horizontal, and otherwise.
We can load hurt triggers that down a player instead of killing them, and also teleport them out of the way.

All these feature require certain entities to already be in the map.  No facilities are provided for adding the required entities to a map, only to use what is already there in a way that makes sense for RotU.