After players win or lose the game, do not make it easy for them to leave the game.  Since players almost unanimously ignore my admonition to stay in the game until the next map loads, disable the menu item that lets them easily leave the game before the next map loads.
This is a server stability issue.  When players stay in the game until the next map loads, we are able to do a full server restart.  However, if there are no players in the game, we can't do a server restart and change the map--we are forced to do a soft reload of the map.  This works ok, for a while.  But if it happens too often in a row, then the CoD4 server runs out of resources and crashes, and will stay down until an admin uses their hosting provider's game control panel to restart the server.

So, wait patiently for the literal minute it takes to vote on a new map and restart the server, and help keep RotU servers running smoothly. :-)