***UNSTABLE***
This commit marks the start of my work to rewrite the bot AI from scratch (informed by the existing code, of course).  The goal is to fix the bugs in the AI, as well as to optimize the AI, waypoints, and spawnpoints.  Until further notice, svn HEAD should be considered unstable, and not fit for use on a production server.  It may still work, but I'm not making any promises. :-)

This commit includes an initial set of hooks to toggle between using the existing AI and my new AI.  The new AI is in bots\bot.gsc and bots\bots.gsc

Features in the new AI in this commit:

* Initial implementation of the Adaptive Slot System, which maximizes available server slots
* Optimizes the getAvailableBot() function from O(n/2) to O(1), where n is the number of bots.
* Optimizes getRandomSpawn() function from O(n) to O(1), where n is the number of entities in the map.
* Finds and adds extra zombie spawnpoints based on the spawnpoints spec'd in the map, to mitigate the issue of zombies spawning inside each other.
* Initial work optimizing bot target selection.