Optimize the waypoints::getNearestWp() function (and rename it to waypoints::nearestWaypoint()).
Over the course of a 15 wave game with seven players, this function is called on the order of 100,000 times. The existing method brute-forced its way through the waypoints array, testing *each* waypoint to find the closest, which of course is O(n) performance per call.

The new method utilizes a k-dimensional tree to find the nearest waypoint, which give O(log n) performance.  This new method only has to examine about 5% as many waypoints as the old brute-force method, resulting in a massively better performance.

Here are my results from two maps:  Boss Road which has a lot of waypoints, and Lake, which is closer to the number of waypoints on a typical map.

For each map, I tested 100,000 random points that were inside the hypercube subtended by the waypoints, and also 100,000 random waypoints that may be inside or outside of the hypercube subtended by the waypoints.

In a real game, players and zombies should be inside the hypercube almost all the time.  The exceptions will be when zombies spawn in areas that aren't waypointed, and when players are near the edge of the map, or have used a map glitch to get out of the waypointed area of the map.  Also, the degradation in performance is relative to how far outside of the hypercube the players or zombies are.  In these tests, the random points that were outside the hypercube could be as much as an additional 50% of the max/min dimension outside the hypercube, which is quite a large distance.  If the player or zombie is only a small distance outside the hypercube, say 5%, the degradation is almost non-existent.

 Starting map Boss Road (mp_surv_boss_road).
 -------------------------------------------------------------------------------
 Waypoint Count: 879 Tree Size: 879
 Tested 100000 random 3D points within the map extents.
 Accuracy (right, wrong): (100000, 0) 100 percent correct.
 Total distance() calls (iterative, kdtree): (87900000, 4453326)
 Average distance() calls (iterative, kdtree): (879, 44.5333)
 Distance Errors (min, max, mean): (0, 0, 0)
 -------------------------------------------------------------------------------
 -------------------------------------------------------------------------------
 Waypoint Count: 879 Tree Size: 879
 Tested 100000 random 3D points.
 Accuracy (right, wrong): (100000, 0) 100 percent correct.
 Total distance() calls (iterative, kdtree): (87900000, 6275633)
 Average distance() calls (iterative, kdtree): (879, 62.7563)
 Distance Errors (min, max, mean): (0, 0, 0)
 -------------------------------------------------------------------------------

 Starting map Lake (mp_lake).
 -------------------------------------------------------------------------------
 Waypoint Count: 438 Tree Size: 438
 Tested 100000 random 3D points within the map extents.
 Accuracy (right, wrong): (100000, 0) 100 percent correct.
 Total distance() calls (iterative, kdtree): (43800000, 2841299)
 Average distance() calls (iterative, kdtree): (438, 28.413)
 Distance Errors (min, max, mean): (0, 0, 0)
 -------------------------------------------------------------------------------
 -------------------------------------------------------------------------------
 Waypoint Count: 438 Tree Size: 438
 Tested 100000 random 3D points.
 Accuracy (right, wrong): (100000, 0) 100 percent correct.
 Total distance() calls (iterative, kdtree): (43800000, 4045112)
 Average distance() calls (iterative, kdtree): (438, 40.4511)
 Distance Errors (min, max, mean): (0, 0, 0)
 -------------------------------------------------------------------------------